pyproject.toml
[tool.poetry]
name = "eval-arena"
version = "0.1.0"
description = ""
authors = ["monami44 <lyaminky2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "0.3.3"
langchain-experimental = "0.3.2"
langchain-huggingface = "0.1.0"
pandas = "2.2.3"
numpy = "^1.26.0"
scikit-learn = "1.5.2"
matplotlib = "3.9.2"
biopython = "1.84"
requests = "2.32.3"
gitpython = "3.1.43"
wikipedia-api = "0.7.1"
secedgar = "0.5.0"
arxiv = "2.1.3"
unittest2 = "1.1.0"
tqdm = "4.66.5"
tenacity = "8.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"